requirements.txt
fastapi
asgi-correlation-id
google-cloud-logging